fix(chat): read a file search's query from pattern, not path
`search_files` carries its query in `pattern`; `path` is the optional
root to search under. The Simple-mode vocabulary read `path`, so it
labelled the search *directory* as the query — a search of `src/billing`
for "invoice" read `Searched files for "billing"` — and dropped the
label entirely for the whole-workspace search that omits `path`, which
is the common case. `target` also picks the axis, contents or filenames,
and calling a filename match "searched files for X" misreports what the
agent looked at.

The branch was unreachable: all three call sites run on
`visibleToolCalls`, which filters `search_files` out. That is exactly
why it was wrong — nothing ever exercised it. `skill_view` crossing
SIMPLE_MODE_HIDDEN_TOOLS in this same branch is the proof that the
policy moves, and a wrong entry ships the moment it does.

So the vocabulary moves out of chat-thread.tsx into simple-labels.ts,
where it can be tested directly instead of only through whichever rows
the current hiding policy happens to render. Pure functions, no React,
no behaviour change; chat-thread.tsx loses ~300 lines and `_toolRowLabel`
loses the underscore now that it has a module boundary to be public at.
